Create 4. Groupby and Orderby
Groupby groups together rows that have the same values in the specified column/s.

GROUPBY allows us to group rows that have the same data and run aggregate functions on them.

When you use group by you have to have the same columns you're grouping on in the group by statement.

The ORDER BY keyword is used to sort the result-set in ascending or descending order.

The ORDER BY keyword sorts the records in ascending order by default. To sort the records in descending order, use the DESC keyword.